feat: Add useAccountToken hook. Remove unnecesary try-catch. Matching component names with file name. Fix message typo. Fix validation logic for manual view
